Move convertHtmlToMarkdown to conversionPipeline.ts
